create function.py
