+++ python/tvm/autotvm/task/relay_integration.py
@@ -61,7 +61,7 @@ def extract_from_program(mod, params, target, target_host=None, ops=None):
 
     Parameters
     ----------
-    mod: tvm.IRModule or relay.expr.Function
+    mod: tvm.IRModule or relay.function.Function
         The module or function to tune
     params: dict of str to numpy array
         The associated parameters of the program
@@ -88,7 +88,7 @@ def extract_from_multiple_program(mods, params, target, target_host=None, ops=No
 
     Parameters
     ----------
-    mods: List[tvm.IRModule] or List[relay.expr.Function]
+    mods: List[tvm.IRModule] or List[relay.function.Function]
         The list of modules or functions to tune
     params: List of dict of str to numpy array
         The associated parameters of the programs
@@ -118,7 +118,7 @@ def extract_from_multiple_program(mods, params, target, target_host=None, ops=No
         logger.disabled = True
 
         for mod, param in zip(mods, params):
-            if isinstance(mod, relay.expr.Function):
+            if isinstance(mod, relay.function.Function):
                 mod = tvm.IRModule.from_expr(mod)
             assert isinstance(mod, tvm.IRModule), \
                 "only support relay Module or Function to be tuned"
