+++ python/tvm/relay/frontend/mxnet.py
@@ -25,6 +25,7 @@ from tvm import relay
 from topi.util import get_const_tuple
 from .. import analysis
 from .. import expr as _expr
+from .. import function as _function
 from .. import op as _op
 from .. import scope_builder as _scope_builder
 from ... import nd as _nd
@@ -1096,7 +1097,7 @@ def _mx_cond(inputs, attrs, subgraphs):
         else_arg_dtype_info = [arg.type_annotation.dtype for arg in else_args]
         else_func = _from_mxnet_impl(subgraphs[2], else_arg_shapes, else_arg_dtype_info)
         sb.ret(_expr.Call(else_func, else_args))
-    func = _expr.Function(input_args, sb.get())
+    func = _function.Function(input_args, sb.get())
     ret = _expr.Call(func, inputs)
     if num_outputs > 1:
         ret = _expr.TupleWrapper(ret, num_outputs)
@@ -1969,7 +1970,7 @@ def _from_mxnet_impl(symbol, shape_dict, dtype_info, params=None, mod=None):
 
     outputs = [node_map[e[0]][e[1]] for e in jgraph["heads"]]
     outputs = outputs[0] if len(outputs) == 1 else _expr.Tuple(outputs)
-    func = _expr.Function(analysis.free_vars(outputs), outputs)
+    func = _function.Function(analysis.free_vars(outputs), outputs)
     return func
